Fetch all the message entries of a particular conversation by passing messageId
Working Flows:

step 1 : First it wil check the requested conversation is exists and current user has allowed to view the conversation.

step 2 : Then, it will check the any new unread message is present in that conversation.

step 3 : If yes, then it will update the conversation(all messages of that conversation) marked as read.

step 4 : And finally, fetch the all messages of that conversation with conversation_status key.

Note: If the conversation was read(which means all messages of that conversation is already read then it will return "Already read" value of conversation_status key.