* Add new keys to show in result like content, participation_mode, etc.
* Modify the existing key name like allDays to all_day.

Note: When we update an entry of a recurring event then that updated entry is saved as new entry. but it has same uid as parent event. So, for this we are adding a key (parent_id) which will be show the id of the original event.

And in 'id' key we are showing the id of updated event entry.