Respect calendar entries over multiple days on start_date filter #67
Now, you will get the event if it is lies in start_date and end_date

For Example: An event is created on 15th sept. and end on 22nd sept. then if you will be fetching the event by passing start_date=2022-09-18 then the event will be getting, because it is lies in the date range